requirements.txt
discord.py
firebase-admin